refactor of managed client
