+++ src/index.ts
@@ -18,23 +18,31 @@
 
 import * as v1 from './v1';
 import * as v1beta1 from './v1beta1';
-import { WriterClient } from './writer_client';
+import * as managedwriter from './managedwriter';
 const BigQueryReadClient = v1.BigQueryReadClient;
 type BigQueryReadClient = v1.BigQueryReadClient;
 const BigQueryWriteClient = v1.BigQueryWriteClient;
 type BigQueryWriteClient = v1.BigQueryWriteClient;
 const BigQueryStorageClient = v1beta1.BigQueryStorageClient;
 type BigQueryStorageClient = v1beta1.BigQueryStorageClient;
+const WriterClient = managedwriter.WriterClient;
+type WriterClient = managedwriter.WriterClient;
 export {
   v1,
   BigQueryReadClient,
   v1beta1,
   BigQueryStorageClient,
   BigQueryWriteClient,
-  WriterClient,
+  managedwriter,
 };
 // For compatibility with JavaScript libraries we need to provide this default export:
 // tslint:disable-next-line no-default-export
-export default {v1, BigQueryReadClient, BigQueryWriteClient, WriterClient};
+export default {
+  v1,
+  managedwriter,
+  BigQueryReadClient,
+  BigQueryWriteClient,
+  WriterClient,
+};
 import * as protos from '../protos/protos';
 export {protos};
