+++ test/managedwriter/writer_client_test.ts
@@ -1,0 +1,50 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// ** This file is automatically generated by gapic-generator-typescript. **
+// ** https://github.com/googleapis/gapic-generator-typescript **
+// ** All changes to this file may be overwritten. **
+
+import * as protos from '../../protos/protos';
+import * as assert from 'assert';
+import * as sinon from 'sinon';
+import {SinonStub} from 'sinon';
+import {describe, it} from 'mocha';
+import {managedwriter} from '../../src';
+
+import {PassThrough} from 'stream';
+
+import {ClientOptions, protobuf} from 'google-gax';
+
+// Dynamically loaded proto JSON is needed to get the type information
+// to fill in default values for request objects
+const root = protobuf.Root.fromJSON(
+  require('../../protos/protos.json')
+).resolveAll();
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+
+describe('v1.WriterClient', () => {
+  describe('Common methods', () => {
+    it('should create a client with the required arguments', () => {
+      const parent =
+        'project/fake-project-id/dataset/fake-dataset-id/table/fake-table-id';
+      const options: ClientOptions = {
+        projectId: 'fake-project-id',
+      };
+      const client = new managedwriter.WriterClient(parent, undefined, options);
+      assert(client);
+    });
+  });
+});
