--- test/veneer/writer_client_test.ts
@@ -1,260 +1,0 @@
-// Copyright 2022 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// ** This file is automatically generated by gapic-generator-typescript. **
-// ** https://github.com/googleapis/gapic-generator-typescript **
-// ** All changes to this file may be overwritten. **
-
-import * as protos from '../../protos/protos';
-import * as assert from 'assert';
-import * as sinon from 'sinon';
-import {SinonStub} from 'sinon';
-import {describe, it} from 'mocha';
-import * as WriterClient from '../../src';
-
-import {PassThrough} from 'stream';
-
-import {ClientOptions, protobuf} from 'google-gax';
-
-// Dynamically loaded proto JSON is needed to get the type information
-// to fill in default values for request objects
-const root = protobuf.Root.fromJSON(
-  require('../../protos/protos.json')
-).resolveAll();
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function getTypeDefaultValue(typeName: string, fields: string[]) {
-  let type = root.lookupType(typeName) as protobuf.Type;
-  for (const field of fields.slice(0, -1)) {
-    type = type.fields[field]?.resolvedType as protobuf.Type;
-  }
-  return type.fields[fields[fields.length - 1]]?.defaultValue;
-}
-
-function generateSampleMessage<T extends object>(instance: T) {
-  const filledObject = (
-    instance.constructor as typeof protobuf.Message
-  ).toObject(instance as protobuf.Message<T>, {defaults: true});
-  return (instance.constructor as typeof protobuf.Message).fromObject(
-    filledObject
-  ) as T;
-}
-
-function stubSimpleCall<ResponseType>(response?: ResponseType, error?: Error) {
-  return error
-    ? sinon.stub().rejects(error)
-    : sinon.stub().resolves([response]);
-}
-
-function stubSimpleCallWithCallback<ResponseType>(
-  response?: ResponseType,
-  error?: Error
-) {
-  return error
-    ? sinon.stub().callsArgWith(2, error)
-    : sinon.stub().callsArgWith(2, null, response);
-}
-
-function stubBidiStreamingCall<ResponseType>(
-  response?: ResponseType,
-  error?: Error
-) {
-  const transformStub = error
-    ? sinon.stub().callsArgWith(2, error)
-    : sinon.stub().callsArgWith(2, null, response);
-  const mockStream = new PassThrough({
-    objectMode: true,
-    transform: transformStub,
-  });
-  return sinon.stub().returns(mockStream);
-}
-
-describe('v1.WriterClient', () => {
-  describe('Common methods', () => {
-    it('has servicePath', () => {
-      const servicePath = WriterClient.WriterClient.servicePath;
-      assert(servicePath);
-    });
-
-    it('has apiEndpoint', () => {
-      const apiEndpoint = WriterClient.WriterClient.apiEndpoint;
-      assert(apiEndpoint);
-    });
-
-    it('has port', () => {
-      const port = WriterClient.WriterClient.port;
-      assert(port);
-      assert(typeof port === 'number');
-    });
-
-    it('should create a client with the required arguments', () => {
-      const parent =
-        'project/fake-project-id/dataset/fake-dataset-id/table/fake-table-id';
-      const options: ClientOptions = {
-        projectId: 'fake-project-id',
-      };
-      const client = new WriterClient.WriterClient(parent, options);
-      assert(client);
-    });
-
-    it('has initialize method and supports deferred initialization', async () => {
-      const parent =
-        'project/fake-project-id/dataset/fake-dataset-id/table/fake-table-id';
-      const opts: ClientOptions = {
-        credentials: {client_email: 'bogus', private_key: 'bogus'},
-        projecId: 'fake-project-id',
-        fallback: true,
-      };
-      const client = new WriterClient.WriterClient(parent, opts);
-      assert.strictEqual(client.writerStub, undefined);
-      await client.initialize();
-      assert(client.writerStub);
-
-      //client initialize methods
-      /*it('should create a client with gRPC fallback', () => {
-      const client = new bigquerywriteModule.v1.BigQueryWriteClient({
-        fallback: true,
-      });
-      assert(client);
-    });*/
-
-      /*it('has initialize method and supports deferred initialization', async () => {
-      const client = new bigquerywriteModule.v1.BigQueryWriteClient({
-        credentials: {client_email: 'bogus', private_key: 'bogus'},
-        projectId: 'bogus',
-      });
-      assert.strictEqual(client.bigQueryWriteStub, undefined);
-      await client.initialize();
-      assert(client.bigQueryWriteStub);
-    });*/
-      // client close methods
-      /*it('has close method for the initialized client', done => {
-      const client = new bigquerywriteModule.v1.BigQueryWriteClient({
-        credentials: {client_email: 'bogus', private_key: 'bogus'},
-        projectId: 'bogus',
-      });
-      client.initialize();
-      assert(client.bigQueryWriteStub);
-      client.close().then(() => {
-        done();
-      });
-    });
-
-    it('has close method for the non-initialized client', done => {
-      const client = new bigquerywriteModule.v1.BigQueryWriteClient({
-        credentials: {client_email: 'bogus', private_key: 'bogus'},
-        projectId: 'bogus',
-      });
-      assert.strictEqual(client.bigQueryWriteStub, undefined);
-      client.close().then(() => {
-        done();
-      });
-    });*/
-
-      // getters and setters
-      it('has getProjectId method', async () => {
-        const fakeParent =
-          'project/fake-project-id/dataset/fake-dataset-id/table/fake-table-id';
-        const fakeProjectId = 'fake-project-id';
-        const fakeOpts: ClientOptions = {
-          credentials: {
-            client_email: 'fake-client@email.com',
-            private_key: 'fake-private-key',
-          },
-          projectId: fakeProjectId,
-        };
-        const client = new WriterClient.WriterClient(fakeParent, fakeOpts);
-        client.auth.getProjectId = sinon.stub().resolves(fakeProjectId);
-        const result = await client.getProjectId();
-        assert.strictEqual(result, fakeProjectId);
-        assert((client.auth.getProjectId as SinonStub).calledWithExactly());
-      });
-
-      it('has getProjectId method with callback', async () => {
-        const fakeParent =
-          'project/fake-project-id/dataset/fake-dataset-id/table/fake-table-id';
-        const fakeProjectId = 'fake-project-id';
-        const fakeOpts: ClientOptions = {
-          credentials: {
-            client_email: 'fake-client@email.com',
-            private_key: 'fake-private-key',
-          },
-          projectId: fakeProjectId,
-        };
-        const client = new WriterClient.WriterClient(fakeParent, fakeOpts);
-        client.auth.getProjectId = sinon
-          .stub()
-          .callsArgWith(0, null, fakeProjectId);
-        const promise = new Promise((resolve, reject) => {
-          client.getProjectId(
-            (err?: Error | null, projectId?: string | null) => {
-              if (err) {
-                reject(err);
-              } else {
-                resolve(projectId);
-              }
-            }
-          );
-        });
-        const result = await promise;
-        assert.strictEqual(result, fakeProjectId);
-      });
-    });
-
-    // Client-Specific Methods
-    /*'getParent',
-      'getStreamId',
-      'createSerializedRows',
-      'getWriteStreams',
-      'setWriteStream',
-      'getWriteStream',
-      'initializeStreamConnection',
-      'appendRowsToStream',
-      'closeStream'*/
-    /*describe('initializeStreamConnection', () => {
-    it('invokes initializeStreamConnection without error', async () => {
-      const fakeParent: string = 'project/fake-project-id/dataset/fake-dataset-id/table/fake-table-id';
-      const fakeProjectId: string = 'fake-project-id';
-      const fakeOpts: ClientOptions = {
-        credentials: {client_email: 'fake-client@email.com', private_key: 'fake-private-key'},
-        projectId: fakeProjectId
-      }
-      const client = new WriterClient.WriterClient(fakeParent, fakeOpts);
-      client.initialize();
-      const request = generateSampleMessage(
-        new protos.google.cloud.bigquery.storage.v1.CreateWriteStreamRequest()
-      );
-      const defaultValue1 = getTypeDefaultValue(
-        '.google.cloud.bigquery.storage.v1.CreateWriteStreamRequest',
-        ['parent']
-      );
-      request.parent = defaultValue1;
-      const expectedHeaderRequestParams = `parent=${defaultValue1}`;
-      const expectedResponse = generateSampleMessage(
-        new protos.google.cloud.bigquery.storage.v1.WriteStream()
-      );
-      client.innerApiCalls.initializeStreamConnection = stubSimpleCall(expectedResponse);
-      const [response] = await client.initializeStreamConnection(request);
-      assert.deepStrictEqual(response, expectedResponse);
-      const actualRequest = (
-        client.innerApiCalls.createWriteStream as SinonStub
-      ).getCall(0).args[0];
-      assert.deepStrictEqual(actualRequest, request);
-      const actualHeaderRequestParams = (
-        client.innerApiCalls.createWriteStream as SinonStub
-      ).getCall(0).args[1].otherArgs.headers['x-goog-request-params'];
-      assert(actualHeaderRequestParams.includes(expectedHeaderRequestParams));
-    });*/
-  });
-});
