+++ tsconfig.json
@@ -7,7 +7,8 @@
     "lib": [
       "es2018",
       "dom"
-    ]
+    ],
+    "module": "CommonJS"
   },
   "include": [
     "src/*.ts",
